Stop stripping quoted text from incoming mail
This was silly, and led to much drama.
Now we're taking the text_part and removing things we don't think we
ought to hand out willy nilly.
Like tokens.
This will make viewing on the web a bit assy, but that's our next story!
There could be a couple integration specs around this, but we need to
have email work on staging and such. Sad times!
[Finishes #47566501]